refactor: move schedule module to lib/zero/schedule/ (#7736)
Part of Epic #7697. Relocates lib/schedule/ into the zero layer at
lib/zero/schedule/ so the zero-run-service import becomes a sibling
reference instead of crossing layer boundaries.
