+++ turbo/apps/web/app/api/cron/execute-schedules/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from "next/server";
 import { initServices } from "../../../../src/lib/init-services";
-import { executeDueSchedules } from "../../../../src/lib/schedule";
+import { executeDueSchedules } from "../../../../src/lib/zero/schedule";
 import { logger } from "../../../../src/lib/logger";
 import { env } from "../../../../src/env";
