+++ turbo/apps/web/app/api/zero/schedules/[name]/enable/route.ts
@@ -5,7 +5,7 @@ import {
   isAuthError,
 } from "../../../../../../src/lib/auth/require-auth";
 import { resolveOrg } from "../../../../../../src/lib/org/resolve-org";
-import { enableSchedule } from "../../../../../../src/lib/schedule";
+import { enableSchedule } from "../../../../../../src/lib/zero/schedule";
 import { isNotFound, isSchedulePast } from "../../../../../../src/lib/errors";
 
 const bodySchema = z.object({
