+++ turbo/apps/web/app/api/zero/schedules/run/route.ts
@@ -3,7 +3,7 @@ import { eq, and } from "drizzle-orm";
 import { initServices } from "../../../../../src/lib/init-services";
 import { getAuthContext } from "../../../../../src/lib/auth/get-auth-context";
 import { resolveOrg } from "../../../../../src/lib/org/resolve-org";
-import { executeSchedule } from "../../../../../src/lib/schedule";
+import { executeSchedule } from "../../../../../src/lib/zero/schedule";
 import { zeroAgentSchedules } from "../../../../../src/db/schema/zero-agent-schedule";
 import { agentRuns } from "../../../../../src/db/schema/agent-run";
